Revert "front end fix"
This reverts commit 7ff04c77fcf03599d53f46114ecd24e75cfa70e1.
